made downloaddaemon deb-friendly and updated install guide

+++ tools/make_version.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
+
+if [ "$1" == "" ]
+then
+	echo "No version number specified"
+	exit
+fi
+
 rm -rf ../version
 
 # ddclient-wx...
-mkdir -p ../version/ddclient-wx-${1}/src/ddclient-wx
+mkdir -p ../version/ddclient-wx-${1}/src/
 mkdir -p ../version/ddclient-wx-${1}/src/lib
 cp -rf ../src/ddclient-wx ../version/ddclient-wx-${1}/src/
 cp -rf ../src/lib/netpptk ../version/ddclient-wx-${1}/src/lib
-cp -rf ../share ../version/ddclient-wx-${1}
+cp -rf ../share ../version/ddclient-wx-${1}/
+cp -f ../AUTHORS ../CHANGES ../TODO ../LICENCE ../INSTALLING ../version/ddclient-wx-${1}/
 echo "cmake_minimum_required (VERSION 2.6)
 
 project(ddclient-wx)
@@ -20,11 +28,21 @@ add_subdirectory(lib)" > ../version/ddclient-wx-${1}/src/CMakeLists.txt
 
 
 
+# DownloadDaemon
+mkdir -p ../version/downloaddaemon-${1}/src/
+cp -rf ../src/daemon ../version/downloaddaemon-${1}/src/
+cp -rf ../src/lib ../version/downloaddaemon-${1}/src/
+cp -rf ../program_data ../version/downloaddaemon-${1}/
+cp -f ../AUTHORS ../CHANGES ../TODO ../LICENCE ../INSTALLING ../version/downloaddaemon-${1}/
+echo "cmake_minimum_required (VERSION 2.6)
 
+project(DownloadDaemon)
+add_subdirectory(src)" > ../version/downloaddaemon-${1}/CMakeLists.txt
 
+echo "cmake_minimum_required (VERSION 2.6)
 
-
-
+project(DownloadDaemon)
+add_subdirectory(daemon)" > ../version/downloaddaemon-${1}/src/CMakeLists.txt
 
 cd ../version
 find -name .svn | xargs rm -rf
